Refactored matchcheck into mvc
